+++ persistence/provider/memory/provider_test.go
@@ -3,7 +3,7 @@ package memory_test
 import (
 	"context"
 
-	"github.com/dogmatiq/infix/persistence/internal/providertest"
+	"github.com/dogmatiq/infix/persistence/provider/internal/providertest"
 	. "github.com/dogmatiq/infix/persistence/provider/memory"
 	. "github.com/onsi/ginkgo"
 )
